requirements.txt
google-api-python-client==2.111.0
anthropic>=0.40.0
resend==0.7.2
python-dotenv==1.0.0
requests>=2.31.0
